backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
joblib
beautifulsoup4
requests
python-dateutil
scipy
